[PISD-59] criada fluxo de cadastro de historico de lesoes. Ajuste de build conforme main
