+++ src/main/java/com/togather/sensei/repositories/HistoricoLesoesRepository.java
@@ -2,6 +2,8 @@ package com.togather.sensei.repositories;
 
 import com.togather.sensei.models.HistoricoLesoesModel;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface HistoricoLesoesRepository extends JpaRepository<HistoricoLesoesModel, Long> {
 }
