+++ src/main/java/com/togather/sensei/services/HistoricoLesoesPostService.java
@@ -1,0 +1,12 @@
+package com.togather.sensei.services;
+
+import com.togather.sensei.models.HistoricoLesoesModel;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public interface HistoricoLesoesPostService {
+
+    HistoricoLesoesModel savehistoricoLesoes(HistoricoLesoesModel historicoLesoesModel);
+}
