+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
-server.port= ${PORT:3000}
+
+server.error.path=/
 
 spring.datasource.url=jdbc:sqlite:identifier.sqlite
 spring.datasource.driver-class-name=org.sqlite.JDBC
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.database-platform=org.hibernate.community.dialect.SQLiteDialect
-
